screens

+++ 824be49b0a23e4664b6156f82ee8ab81/Models/SpaceStationModel.swift
@@ -10,8 +10,8 @@ import Foundation
 // MARK: - SpaceStationModelElement
 struct SpaceStationModelElement: Codable {
     let name: String
-    let coordinateX: Int
-    let coordinateY: Int
+    let coordinateX: Double
+    let coordinateY: Double
     let capacity: Int
     let stock: Int
     let need: Int
